pyproject.toml
[project]
name = "entiendo"
version = "0.1.0"
description = "AI-powered document understanding with citations"
readme = "README.md"
authors = [
    {name = "Joshua Oliphant", email = "joshuaoliphant@hey.com"},
]
dependencies = [
    "anthropic>=0.45.2",
    "fastapi>=0.115.7",
    "jinja2>=3.1.5",
    "logfire[fastapi]>=3.4.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
